Refactoring and additional features:
Changed the way we use types to define enrichments.
Changed the way the enrichment factory works accordingly to the types.
Changed swagger definition and updated the swagger docs.
Changed the way we generate ID's, we now use MD5 hashing.
Changed the way we consume messages from kafka to use a consumer stream
Changes in logic of how to handle the reading and writing to kafka